+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/types/BTypedescType.java
@@ -58,7 +58,7 @@ public class BTypedescType extends BType {
             return true;
         }
         if (obj instanceof BTypedescType) {
-            return constraint == ((BTypedescType) obj).getConstraint();
+            return constraint.equals(((BTypedescType) obj).getConstraint());
         }
         return false;
     }
