+++ tests/jballerina-unit-test/src/test/java/org/ballerinalang/test/statements/ifelse/TypeGuardTest.java
@@ -598,4 +598,16 @@ public class TypeGuardTest {
     public void testSameVarNameInDifferentScopes() {
         BRunUtil.invoke(result, "testSameVarNameInDifferentScopes");
     }
+
+    @Test(description = "Test Typetest for TypeDefs when types are equal")
+    public void testTypetestForTypedefs1() {
+        BValue[] returns = BRunUtil.invoke(result, "testTypeDescTypeTest1");
+        Assert.assertEquals(BBoolean.TRUE, (BBoolean) returns[0]);
+    }
+
+    @Test(description = "Test Typetest for TypeDefs when types are not equal")
+    public void testTypetestForTypedefs2() {
+        BValue[] returns = BRunUtil.invoke(result, "testTypeDescTypeTest2");
+        Assert.assertEquals(BBoolean.TRUE, (BBoolean) returns[0]);
+    }
 }
